Actually fixed the issue with a cover called cover.jpg

+++ assets/php/add_album_finalize.php
@@ -41,7 +41,8 @@ if (mysqli_error($mysqli)) {
         unset($_SESSION['covers'][$picked_cover]);
 
         foreach ($_SESSION['covers'] as $cover) {
-            unlink($tmp_folder . $cover);
+            if ($cover != "cover.jpg")
+                unlink($tmp_folder . $cover);
         }
 
         session_destroy();
